refactor: add auth module. move login and register to auth
