+++ app/src/main/java/ro/code4/deurgenta/ui/auth/AuthViewModel.kt
@@ -1,4 +1,4 @@
-package ro.code4.deurgenta.ui.login
+package ro.code4.deurgenta.ui.auth
 
 import android.content.SharedPreferences
 import androidx.lifecycle.LiveData
@@ -6,16 +6,21 @@ import org.koin.core.inject
 import ro.code4.deurgenta.helper.Result
 import ro.code4.deurgenta.helper.SingleLiveEvent
 import ro.code4.deurgenta.helper.hasCompletedOnboarding
+import ro.code4.deurgenta.repositories.Repository
 import ro.code4.deurgenta.ui.base.BaseViewModel
 import ro.code4.deurgenta.ui.main.MainActivity
 import ro.code4.deurgenta.ui.onboarding.OnboardingActivity
 
-class LoginFormViewModel : BaseViewModel() {
+class AuthViewModel : BaseViewModel() {
+
+    private val repository: Repository by inject()
     private val sharedPreferences: SharedPreferences by inject()
 
     private val loginLiveData = SingleLiveEvent<Result<Class<*>>>()
-    fun loggedIn(): LiveData<Result<Class<*>>> = loginLiveData
+    private val registerLiveData = SingleLiveEvent<Result<Class<*>>>()
 
+    fun loggedIn(): LiveData<Result<Class<*>>> = loginLiveData
+    fun registered(): LiveData<Result<Class<*>>> = registerLiveData
 
     fun login() {
         val nextActivity = when (sharedPreferences.hasCompletedOnboarding()) {
@@ -24,4 +29,13 @@ class LoginFormViewModel : BaseViewModel() {
         }
         loginLiveData.postValue(Result.Success(nextActivity))
     }
+
+    fun onRegisterSuccess() {
+        registerLiveData.postValue(Result.Success())
+    }
+
+    fun onRegisterFail(error: Throwable, message: String = "") {
+        registerLiveData.postValue(Result.Failure(error, message))
+    }
+
 }
