+++ app/src/main/res/layout-land/activity_auth.xml
@@ -7,7 +7,7 @@
     android:clipChildren="false"
     android:clipToPadding="false"
     android:padding="@dimen/margin"
-    tools:context="ro.code4.deurgenta.ui.login.LoginActivity">
+    tools:context="ro.code4.deurgenta.ui.auth.AuthActivity">
 
     <ImageView
         android:id="@+id/logo"
