+++ app/src/main/res/layout/activity_auth.xml
@@ -6,7 +6,7 @@
     android:clipChildren="false"
     android:id="@+id/login_container"
     android:clipToPadding="false"
-    tools:context="ro.code4.deurgenta.ui.login.LoginActivity">
+    tools:context="ro.code4.deurgenta.ui.auth.AuthActivity">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
