+++ app/src/main/res/layout/fragment_completed_register.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         xmlns:tools="http://schemas.android.com/tools"
-        tools:context="ro.code4.deurgenta.ui.login.LoginActivity"
+        tools:context="ro.code4.deurgenta.ui.auth.AuthActivity"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/colorPrimaryLight"
