+++ app/src/main/res/layout/fragment_register.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
-        tools:context="ro.code4.deurgenta.ui.login.LoginActivity"
+        tools:context="ro.code4.deurgenta.ui.auth.AuthActivity"
         xmlns:app="http://schemas.android.com/apk/res-auto">
     <data>
         <import type="android.view.View"/>
         <variable
                 name="viewmodel"
-                type="ro.code4.deurgenta.ui.register.RegisterViewModel" />
+                type="ro.code4.deurgenta.ui.auth.register.RegisterViewModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
